Input stuff moved to Raise project. ReferenceCounter for resources. CShell is being implemented. GConsole stuff introduced. VAnimation status event implemented. VDraw new functions for rendering TRegions. Some VGUI class imrpovements for support multiple desktops. VScene::Flags and Type introduced. LayeredScene renderer and basic pipeline like rendering classes  implemented.

+++ realizer/vnodecanvas.cpp
@@ -14,8 +14,8 @@ VNodeCanvasImage* VNodeCanvas::DrawImage( VTexture* Texture, int dstX,int dstY, 
 	img->Texture = Texture;
 	img->Part.Initialize(Texture,0,0,Texture->Width,Texture->Height);
 
-	img->Position.x = dstX;
-	img->Position.y = dstY;
+	img->Position.x = (float)dstX;
+	img->Position.y = (float)dstY;
 
 	img->Color.color = 0xFFFFFFFF;
